Structure refactor
Existing classes are moved to Utils package
