+++ src/main/java/challenge/mutantes/Utils/DnaAnalyzer.java
@@ -1,4 +1,4 @@
-package challenge.mutantes;
+package challenge.mutantes.Utils;
 
 import java.util.ArrayList;
 import java.util.List;
