+++ src/main/java/challenge/mutantes/Utils/HorizontalFinder.java
@@ -1,4 +1,4 @@
-package challenge.mutantes;
+package challenge.mutantes.Utils;
 
 public class HorizontalFinder extends Finder {
     public HorizontalFinder(Point point) {
