+++ src/main/java/challenge/mutantes/Utils/Point.java
@@ -1,4 +1,4 @@
-package challenge.mutantes;
+package challenge.mutantes.Utils;
 
 import java.util.Objects;
