+++ src/main/java/challenge/mutantes/Utils/RightDiagonalFinder.java
@@ -1,4 +1,4 @@
-package challenge.mutantes;
+package challenge.mutantes.Utils;
 
 public class RightDiagonalFinder extends Finder {
     public RightDiagonalFinder(Point point) {
