+++ src/test/java/challenge/mutantes/Utils/HorizontalFinderTest.java
@@ -1,5 +1,7 @@
-package challenge.mutantes;
+package challenge.mutantes.Utils;
 
+import challenge.mutantes.Utils.HorizontalFinder;
+import challenge.mutantes.Utils.Point;
 import org.junit.Test;
 
 import java.util.ArrayList;
@@ -9,7 +11,7 @@ import static org.junit.Assert.*;
 public class HorizontalFinderTest {
     @Test public void testGeneratedPoints() {
         HorizontalFinder finder = new HorizontalFinder(new Point(0, 0));
-        ArrayList<challenge.mutantes.Point> generatedPoints = finder.getPoints();
+        ArrayList<Point> generatedPoints = finder.getPoints();
 
         assertEquals(new Point(1, 0), generatedPoints.get(0));
         assertEquals(new Point(2, 0), generatedPoints.get(1));
