+++ src/test/java/challenge/mutantes/Utils/RightDiagonalFinderTest.java
@@ -1,5 +1,7 @@
-package challenge.mutantes;
+package challenge.mutantes.Utils;
 
+import challenge.mutantes.Utils.Point;
+import challenge.mutantes.Utils.RightDiagonalFinder;
 import org.junit.Test;
 
 import java.util.ArrayList;
